smarthome-web-portal/tools/smarthome-admin-portal/requirements.txt
anyjson==0.3.3
billiard==3.3.0.22
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.8
kombu==3.0.30
MarkupSafe==0.23
PyMySQL==0.7.6
pytz==2015.7
six==1.10.0
Werkzeug==0.11
Flask-SQLAlchemy==2.1
numpy
scipy
matplotlib
